requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.0.1
watchdog==6.0.0
flask-sock==0.7.0